backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
python-multipart==0.0.6
fastapi-cors==0.0.6
websockets==12.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2